Add Day 1 Part 2 with Gleam
As always, algorithm in the docs. main was pretty nicely chunked down
thanks to Gleam being pretty functional, so we only had to change a
couple of things in the pipeline.

Moving forward: I want to check real-world Gleam code how they tend to
deal with Errors, especially around main(). let assert will crash. Is
that the right thing? It maps to exit code 2.

Perhaps if we're writing a Unix program, I'm meant to return a nicer
exit code and print to stderr, but how do I do those things? That'd be
nice follow-up reading.