Add Day 2 Part 1 with Ocaml
I've always wanted to try! The onboarding experience is pretty modern in
a lot of ways. Opam and Dune help keep the development environment
tooled cohesively (pretty telling mark of a modern language), the
quick-start guide in the docs allows one to jump right in, etc.

This was easier for me to pick up compared to Gleam, probably because
- I didn't realize Gleam was functional in the first place! I had no
  idea what to expect
- I can *feel* that Ocaml and Haskell are related. I was able to apply
  patterns I already knew

As always, the algorithm is in the docs. One notable detour was I
briefly considered an iterative version of is_report_safe which used
List.mapi.  I moved away from it because I couldn't build in the
short-circuit clause (as long as a single check fails, I can just fail
the entire report!).

I've also added compilation to the Makefile, but it's actually
unnecessary (dune exec will automatically compile if a build doesn't
exist anyway). I never really looked into whether this was how gleam run
worked, so I might take a look back.